#Disconnect the login from the rest of the CRUD  methods for user.
